Utveckling av webbapplikationer/Flask/Tutorial/secondo tentativo/requirements.txt
email-validator
flask_wtf
wtforms
uuid
flask-login